language/totto/eval_requirements.txt
absl-py
pytest
sacrebleu
six
wheel
